requirements.txt
Django~=5.1
django-environ==0.11.2
django-simple-search==1.0.2
Markdown==3.7
rich==13.9.4
pytz==2024.2
python-slugify==8.0.4
Jinja2==3.1.4
